requirements.txt
Django==5.0.6
djangorestframework==3.15.1
mysqlclient==2.2.4
django-mysql-pool==0.2
dj-database-url==2.2.0
python-dotenv==1.0.1
gunicorn==22.0.0
whitenoise==6.7.0
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
celery==5.4.0
redis==5.0.5
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
requests==2.32.3
django-storages[boto3]==1.14.2
boto3==1.34.140
google-generativeai==0.7.1
drf-spectacular==0.27.2
flake8==7.1.0
Pygments==2.17.2
django-ratelimit==4.0.0
sentry-sdk[django]==2.0.1
bleach==6.1.0
django-celery-beat==2.6.0
django-ses==3.2.0
django-redis==5.4.0
